refactor(game): remove CreatureDbToCreatureConverter
