+++ src/data-modules/game/adapter/config/container/adapter/bindGameAdapterDb.ts
@@ -1,9 +1,8 @@
 import { interfaces } from 'inversify';
 
+import { CardCreationQueryToCardDbsConverter } from '../../../db/converter/card/CardCreationQueryToCardDbsConverter';
 import { CardDbToCardConverter } from '../../../db/converter/card/CardDbToCardConverter';
 import { CardFindQueryToCardDbFilterQueryConverter } from '../../../db/converter/card/CardFindQueryToCardDbFilterQueryConverter';
-import { CardCreationQueryToCardDbsConverter } from '../../../db/converter/card/CreatureCreationQueryToCreatureDbsConverter';
-import { CreatureDbToCreatureConverter } from '../../../db/converter/card/CreatureDbToCreatureConverter';
 import { CardDeckCreationQueryToCardDeckDbsConverter } from '../../../db/converter/deck/CardDeckCreationQueryToCardDeckDbsConverter';
 import { CardDeckDbToCardDeckConverter } from '../../../db/converter/deck/CardDeckDbToCardDeckConverter';
 import { CardDeckFindQueryToCardDeckDbFilterQueryConverter } from '../../../db/converter/deck/CardDeckFindQueryToCardDeckDbFilterQueryConverter';
@@ -57,9 +56,6 @@ export function bindGameAdapterDb(bind: interfaces.Bind): void {
     GAME_ADAPTER_TYPES.db.converter.card
       .CARD_CREATION_QUERY_TO_CARD_DBS_CONVERTER,
   ).to(CardCreationQueryToCardDbsConverter);
-  bind(
-    GAME_ADAPTER_TYPES.db.converter.card.CREATURE_DB_TO_CREATURE_CONVERTER,
-  ).to(CreatureDbToCreatureConverter);
 
   bind(
     GAME_ADAPTER_TYPES.db.converter.deck
