+++ src/data-modules/game/adapter/config/types/gameAdapterDbTypes.ts
@@ -21,9 +21,6 @@ export const GAME_ADAPTER_DB_TYPES = {
       CARD_CREATION_QUERY_TO_CARD_DBS_CONVERTER: Symbol(
         'CreatureCreationQueryToCreatureDbsConverter',
       ),
-      CREATURE_DB_TO_CREATURE_CONVERTER: Symbol(
-        'CreatureDbToCreatureConverter',
-      ),
     },
     deck: {
       CARD_DECK_CREATION_QUERY_TO_CARD_DBS_CONVERTER: Symbol(
